Bug fix for displaying accurate temperature on WIO Terminal and MQTT
imported a library called DHT.h to connect the Grove Temp & Humidity sensor for detecting the correct temperature values in celsius. The temperature readings were double checked with a thermostat. This is an improvement from previous code, as there are fewer temperature fluctuations and the temperature is closer to the the thermostat readingsCo-authored By: Vaibhav Puram puram@net.chalmers.se