#25 Improve UI intuitiveness
As a result of customer feedback during a live demo of the product in progress which states that it is difficult to understand the purpose of the screen until it is explained, a label prompting the user what exactly to do is added around the top of the screen. Additionally, the 'Threshold' label next to the slider temperature display is changed with 'Switch at' to make it clear what exactly setting the temperature does.