#25 Fix confirmation message bug
The confirmation message was not popping up but was instead throwing an error due to the animated view for the fade in of the message. The animation was removed as it was recognized to serve no purpose in the display of the confirmation message, and instead make the development slower.