#25 Modify addThreshold to updateThreshold
The addThreshold method was adding a new entry to the firebase everytime the user was saving the threshold values. This became a problem in the attempt to actually then find the latest entry and display it when the app is restarted. Since there are no login details connected to each user,it seems easier to just update the same entry everytime values are changed. However, the functionality of displaying the previously set values when the app is restarted is yet to be added. The current change just makes this process more convinient and allows for the removal of unnecessary code that can be difficult to maintain.