#25 Segregate MQTT responsibility
This is done to facilitate the publishing of temperatureThreshold values to the MQTT broker. The previous structure required a complex approach where the thresholds would have to be passed from a child to parent component. Seperating mqtt into its own class, however, supports single responsibility and can allow for the values to be published with just an import and a function call in TemperatureThresholds.js. This is yet to be done.