#25 Successfully publish values to MQTT broker
The publishToTopic method in mqttUtils had to be modified as client.publish was not a suitable method here. Instead the method now creates a message object of the thresholds set by the user and sends it to the broker to the mentioned topic. The publishToTopic method has been made highly reusable such that any message to any topic can be published. The updateThreshold method connects to the client when settings are saved and calls on this publishToTopic method.