#5 Implement mqttUtils for publishing temperature units
The mqttUtils.js file was copied from branch #25 so that the temperature units can be published directly from the Metric.js file. This will decrease the coupling between classes regarding the responsibility for MQTT publishing/ subscribing. Moreover, it will avoid sharing the metricUnit state between App component and MetricDisplayWidget components and publish the values directly from the Metric component to the MQTT broker. This ultimately keeps the responsibility of publishing temoerature units to the MQTT broker to the Metric component.