#34 Change the Condition the Mode is Highlighted to track the selectedModeId instead
Previously the condition to highlight the mode was directly via the mode's stored selection boolean in the databse, which in turn allowed the user with fast clicks on multiple moods to highlight more than one mode at same time since all modes selected fasted was meeting the condition at evaluation. However, with evaluating current selectedModeId's stat if it is equal to mode's real db id, eliminates this issue.