#31 Store default constants in seperate class
As the default checkbox values were hard coded and can make it difficult for future maintainability if other classes were to use the same variable, the constants are now kept in a seperate class called LogicConstants under the constants folder so any component can access them. This increase maintainability and readibility.