#20 Implement Hook to Display and Store Energy at an Interval of 1 second and reset state to updated level
This Hook Checks if it is a custom or auto mode and based on that it eiter covnert the auto's dutycycles received to normal fan speed or if it is a custom mode it will get it's fan speed and in both cases it will use energyCalculator to calculate energy consumed and store it in firebase at an interval of one second. The hook updates if the autoDutyCyle, fanIsOn, modes or the current selectedModeId change.