#35 Decouple temperature unit change and temperature value change upon button press in phone app
This change provides a bug-fix for the temperature reading reverting back to celsius whenever a change in unit is triggered. This also enables the temperature reading on both the phone app and WIO Terminal to update faster. Initially, the convertTemperature() function had the responsibility of changing both the temperature reading and unit, however subscribed temperature from the WIO Terminal was still in degree celsuis. Now, the convertSubscribedTemperature() function handles subscribed temperature from the WIO and converts it according to the selected unit on the app for an accurate in-app reading